업데이트

+++ HR-main/src/main/java/main/DeductionResult.java
@@ -1,24 +1,22 @@
 package main;
 
-// 계산된 모든 공제 항목을 담는 데이터 클래스
+import java.math.BigDecimal;
+
 public class DeductionResult {
-    // 근로자 부담분
-    double nationalPensionEmployee;
-    double healthInsuranceEmployee;
-    double longTermCareInsuranceEmployee;
-    double employmentInsuranceEmployee;
-    double incomeTax;
-    double localIncomeTax;
-    double totalEmployeeDeduction;
+    BigDecimal nationalPensionEmployee;
+    BigDecimal healthInsuranceEmployee;
+    BigDecimal longTermCareInsuranceEmployee;
+    BigDecimal employmentInsuranceEmployee;
+    BigDecimal incomeTax;
+    BigDecimal localIncomeTax;
+    BigDecimal totalEmployeeDeduction;
 
-    // 사업주 부담분
-    double nationalPensionEmployer;
-    double healthInsuranceEmployer;
-    double longTermCareInsuranceEmployer;
-    double employmentInsuranceEmployer;
-    double industrialAccidentInsuranceEmployer;
-    double totalEmployerDeduction;
+    BigDecimal nationalPensionEmployer;
+    BigDecimal healthInsuranceEmployer;
+    BigDecimal longTermCareInsuranceEmployer;
+    BigDecimal employmentInsuranceEmployer;
+    BigDecimal industrialAccidentInsuranceEmployer;
+    BigDecimal totalEmployerDeduction;
 
-    // 실지급액
-    double netPay;
+    BigDecimal netPay;
 }
